move mapCss to separate module for easier isolated testing
